requirements.txt
fastapi[all]
uvicorn
sqlalchemy
asyncpg
python-dotenv
pydantic
bcrypt
alembic